requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
openai>=1.30.0
google-generativeai>=0.7.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.130.0
psycopg2-binary>=2.9.9
fastapi>=0.111.0
uvicorn>=0.29.0
apscheduler>=3.10.4
python-dotenv>=1.0.1
pydantic>=2.7.0
